Tela funcional

+++ include/Jogador.hpp
@@ -11,7 +11,7 @@ namespace Personagens {
         bool joga1;
         // bool noChao;
     public:
-        Jogador(const int x_inicial = 0, const int y_inicial = 0);
+        Jogador(const float x_inicial = 0, const float y_inicial = 0);
         ~Jogador();
 
         const int getPontos() const { return pontos; }
